flask_app/requirements.txt
langchain
langchain-community
openai
tiktoken
pinecone-client
langchain_pinecone
unstructured
pdfminer==20191125
pdfminer.six==20221105
pillow_heif
unstructured_inference
youtube-transcript-api
pytube
sentence-transformers
flask